feat(auth): refactor forget password flow with new BLoC structure and remove deprecated files
