+++ lib/features/auth/presentation/view_model/forget_password_view_model/forget_password_cubit.dart
@@ -1,8 +1,8 @@
 import 'package:exam_app/confing/api_result/api_result.dart';
 import 'package:exam_app/features/auth/domin/entities/forgot_password_req.dart';
 import 'package:exam_app/features/auth/domin/usecases/forget_password_usecase.dart';
-import 'package:exam_app/features/auth/presentation/forget_password_events.dart';
-import 'package:exam_app/features/auth/presentation/forget_password_state.dart';
+import 'package:exam_app/features/auth/presentation/view_model/forget_password_view_model/forget_password_events.dart';
+import 'package:exam_app/features/auth/presentation/view_model/forget_password_view_model/forget_password_state.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:injectable/injectable.dart';
 @injectable
