--- lib/features/auth/presentation/view_model/reset_password_view_model/reset_password_view_model.dart
@@ -1,55 +1,0 @@
-import 'package:exam_app/confing/api_result/api_result.dart';
-
-import 'package:exam_app/features/auth/domin/usecases/reset_password_usecase.dart';
-
-import 'package:exam_app/features/auth/presentation/view_model/reset_password_view_model/reset_password_events.dart';
-import 'package:exam_app/features/auth/presentation/view_model/reset_password_view_model/reset_password_states.dart';
-import 'package:flutter/material.dart';
-import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:injectable/injectable.dart';
-
-@injectable
-class ResetPasswordViewModel
-    extends Bloc<ResetPasswordEvents, ResetPasswordStates> {
-  final ResetPasswordUsecase _resetPasswordUsecase;
-  final GlobalKey<FormState> formKey = GlobalKey();
-
-  ResetPasswordViewModel(this._resetPasswordUsecase)
-    : super(ResetPasswordStates()) {
-    on<ResetNewPasswordEvent>(_resetNewPassword);
-    on<ValidateResetPasswordEvent>(_checkValidation);
-  }
-
-  void _checkValidation(ValidateResetPasswordEvent event, Emitter emit) {
-    final valid = formKey.currentState!.validate();
-
-
-    emit(state.copyWith(isValid: valid));
-  }
-
-  Future<void> _resetNewPassword(
-    ResetNewPasswordEvent event,
-    Emitter emit,
-  ) async {
-    emit(state.copyWith(isLoading: true));
-
-    final res = await _resetPasswordUsecase.resetPassword(
-      event.resetPasswordRequest,
-    );
-    switch (res) {
-      case ApiSucessResult<void>():
-        emit(state.copyWith(isLoading: false, isSuccess: true));
-
-      case ApiFailedResult<void>():
-        emit(
-          state.copyWith(
-            isLoading: false,
-            errorMessage: res.errorMessage,
-            isSuccess: false,
-          ),
-        );
-    }
-
-    //
-  }
-}
