--- lib/features/auth/presentation/view_model/verify_code_view_model/verify_code_state.dart
@@ -1,27 +1,0 @@
-class VerifyCodeState {
-  bool isLoading;
-  String? errorMessage;
-  bool isSuccess;
-  bool? isCodeResend;
-
-  VerifyCodeState({
-    this.isLoading = false,
-    this.errorMessage,
-    this.isSuccess = false,
-    this.isCodeResend=false
-  });
-
-  VerifyCodeState copyWith({
-    bool? isLoading,
-    String? errorMessage,
-
-    bool? isSuccess,
-  }) {
-    return VerifyCodeState(
-      isLoading: isLoading ?? this.isLoading,
-      errorMessage: errorMessage ?? this.errorMessage,
-
-      isSuccess: isSuccess ?? this.isSuccess,
-    );
-  }
-}
