+++ lib/features/auth/presentation/views/widgets/register_button.dart
@@ -1,18 +1,12 @@
 import 'package:exam_app/core/theme/app_colors.dart';
-import 'package:exam_app/features/auth/domin/entities/forgot_password_req.dart';
-import 'package:exam_app/features/auth/domin/entities/reset_password_req.dart';
+
 import 'package:exam_app/features/auth/domin/entities/signin_req.dart';
 import 'package:exam_app/features/auth/domin/entities/signup_req.dart';
 
-import 'package:exam_app/features/auth/presentation/view_model/forgot_password_view_model/forgot_password_events.dart';
-import 'package:exam_app/features/auth/presentation/view_model/forgot_password_view_model/forgot_password_view_model.dart';
-import 'package:exam_app/features/auth/presentation/view_model/reset_password_view_model/reset_password_events.dart';
-import 'package:exam_app/features/auth/presentation/view_model/reset_password_view_model/reset_password_view_model.dart';
 import 'package:exam_app/features/auth/presentation/view_model/signin_cubit/signin_cubit.dart';
 
 import 'package:exam_app/features/auth/presentation/view_model/signup_view_model/signup_events.dart';
 import 'package:exam_app/features/auth/presentation/view_model/signup_view_model/signup_view_model.dart';
-
 
 import 'package:flutter/material.dart';
 
@@ -26,12 +20,12 @@ class RegisterButton extends StatelessWidget {
     this.signInRequest,
     this.signinCubit,
     this.text,
-    this.forgotPasswordViewModel,
-    this.forgotPasswordRequest,
+    // this.forgotPasswordViewModel,
+    // this.forgotPasswordRequest,
     this.controller1,
     this.controller2,
-    this.resetPasswordRequest,
-    this.resetPasswordViewModel,
+    // this.resetPasswordRequest,
+    // this.resetPasswordViewModel,
   });
 
   final bool active;
@@ -41,17 +35,16 @@ class RegisterButton extends StatelessWidget {
   final SigninCubit? signinCubit;
   final SignupRequest? signupRequest;
   final SignInRequest? signInRequest;
-  final ForgotPasswordViewModel? forgotPasswordViewModel;
-  final ForgotPasswordRequset? forgotPasswordRequest;
-  final ResetPasswordRequest? resetPasswordRequest;
-  final ResetPasswordViewModel? resetPasswordViewModel;
+  // final ForgotPasswordViewModel? forgotPasswordViewModel;
+  // final ForgotPasswordRequset? forgotPasswordRequest;
+  // final ResetPasswordRequest? resetPasswordRequest;
+  // final ResetPasswordViewModel? resetPasswordViewModel;
 
   final TextEditingController? controller1;
   final TextEditingController? controller2;
 
   @override
   Widget build(BuildContext context) {
-   
     return ElevatedButton(
       style: ButtonStyle(
         elevation: WidgetStateProperty.all(active ? 10 : 0),
@@ -64,41 +57,36 @@ class RegisterButton extends StatelessWidget {
         ),
       ),
       onPressed: () {
-       
-
         if (active) {
           if (signupViewModel != null && signupRequest != null) {
             signupViewModel!.add(SignUpUserEvent(signupRequest!));
-          }
-           else if (signinCubit != null && signInRequest != null) {
+          } else if (signinCubit != null && signInRequest != null) {
             if (signinCubit!.checkValidation()) {
               signinCubit!.signIn(signInRequest!);
             }
           } 
-          else if (forgotPasswordViewModel != null &&forgotPasswordRequest != null) {
-            forgotPasswordViewModel!.add(
-              SendForgotPasswordEmailEvent(
-                ForgotPasswordRequset(email: controller1!.text.trim()),
-              ),
-            );
-          }
-
-           else if (
-            resetPasswordViewModel != null) {
           
-
-          resetPasswordViewModel!.add(
-            ResetNewPasswordEvent(
-              ResetPasswordRequest(
-                controller1!.text.trim(),
-                controller2!.text.trim(),
-              ),
-            ),
-          );
+          
+          // else if (forgotPasswordViewModel != null &&
+          //     forgotPasswordRequest != null) {
+          //   forgotPasswordViewModel!.add(
+          //     SendForgotPasswordEmailEvent(
+          //       ForgotPasswordRequset(email: controller1!.text.trim()),
+          //     ),
+          //   );
+          // } 
+          
+          // else if (resetPasswordViewModel != null) {
+          //   resetPasswordViewModel!.add(
+          //     ResetNewPasswordEvent(
+          //       ResetPasswordRequest(
+          //         controller1!.text.trim(),
+          //         controller2!.text.trim(),
+          //       ),
+          //     ),
+          //   );
+          // }
         }
-        } 
-        
-       
       },
 
       child: Padding(
